working on spinner

+++ src/animation/mod.rs
@@ -23,7 +23,7 @@ pub enum Easing {
 }
 
 impl Easing {
-    pub fn interpolate(&self, progress: f64) -> f64 {
+    pub fn interpolate(&self, progress: f32) -> f32 {
         assert!(progress >= 0.0 && progress <= 1.0);
         match self {
             Self::Linear => progress,
@@ -32,5 +32,5 @@ impl Easing {
 }
 
 pub trait Interpolateable {
-    fn interpolate(start: &Self, end: &Self, progress: f64) -> Self;
+    fn interpolate(start: &Self, end: &Self, progress: f32) -> Self;
 }
